refactored services to remove extra assembly
