--- Metadata.Xml/Internals.cs
@@ -1,6 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Runtime.CompilerServices;
-using System.Text;
-
-[assembly: InternalsVisibleTo("StateChartsDotNet.Services")]
