+++ Metadata.Xml/Services/Resolver.cs
@@ -1,14 +1,13 @@
 ﻿using StateChartsDotNet.Common;
 using StateChartsDotNet.Metadata.Xml.Execution;
-using StateChartsDotNet.Services.Http;
 using System;
 using System.Collections.Generic;
 using System.Text;
 using System.Xml.Linq;
 
-namespace StateChartsDotNet.Services
+namespace StateChartsDotNet.Metadata.Xml.Services
 {
-    public static class XmlResolver
+    internal static class Resolver
     {
         public static SendMessageMetadata Resolve(XElement element)
         {
@@ -16,7 +15,7 @@ namespace StateChartsDotNet.Services
             
             SendMessageMetadata metadata = element.Name.LocalName switch
             {
-                "http-post" => new XmlMetadata(element),
+                "http-post" => new HttpPostMetadata(element),
                 _ => throw new NotSupportedException("Unexpected service type: " + element.Name.LocalName),
             };
