+++ Metadata.Xml/States/StateChart.cs
@@ -12,8 +12,6 @@ namespace StateChartsDotNet.Metadata.Xml.States
 {
     public class StateChart : StateMetadata, IRootStateMetadata
     {
-        public static Func<XElement, SendMessageMetadata> ServiceResolver = null;
-
         public StateChart(XDocument document)
             : base(document.Root)
         {
