--- Services/Services.csproj
@@ -1,20 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-    <AssemblyName>StateChartsDotNet.Services</AssemblyName>
-    <RootNamespace>StateChartsDotNet.Services</RootNamespace>
-    <LangVersion>9.0</LangVersion>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Common\Common.csproj" />
-    <ProjectReference Include="..\Metadata.Fluent\Metadata.Fluent.csproj" />
-    <ProjectReference Include="..\Metadata.Xml\Metadata.Xml.csproj" />
-  </ItemGroup>
-
-</Project>
