+++ StateChartsDotNet.sln
@@ -20,8 +20,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Metadata.Xml", "Metadata.Xm
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Tests", "Tests\Tests.csproj", "{898762AE-3EE2-45DD-8E88-CD7E958BFACC}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Services", "Services\Services.csproj", "{FB339B24-C6F8-4657-AB09-7F27F3ED9759}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -52,10 +50,6 @@ Global
 		{898762AE-3EE2-45DD-8E88-CD7E958BFACC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{898762AE-3EE2-45DD-8E88-CD7E958BFACC}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{898762AE-3EE2-45DD-8E88-CD7E958BFACC}.Release|Any CPU.Build.0 = Release|Any CPU
-		{FB339B24-C6F8-4657-AB09-7F27F3ED9759}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{FB339B24-C6F8-4657-AB09-7F27F3ED9759}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{FB339B24-C6F8-4657-AB09-7F27F3ED9759}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{FB339B24-C6F8-4657-AB09-7F27F3ED9759}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
